Agent Creator + Terrain Map clean up + Falloff Map
Agent Creator Update:
I've somewhat implemented bones however it's not exactly working as intended. Initially if a bone was altered it would affect the vertices bound to it but not the whole mesh, I originally thought it had to do with the DestroyImmediate function in the AddBone class as it was not activating due to the fact that function doesn't work in editor mode therefore I decided to implement a HUD so I can test it in play mode but when I did this it broke the whole system and bones don't work at all now. I'm going to leave this as is for now and come back to it.

Terrain Map Update: I decided to clean up the terrain generation code in the main MASE project.

Falloff map: I made a falloff map and applied it to the terrain map so that I could essentially create "Islands"